Rename: Guide파일 경로 수정(common -> pages)
