+++ src/pages/Guide.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import { Button } from "./Button";
-import { Input } from "./Input";
+import { Button } from "../components/common/Button";
+import { Input } from "../components/common/Input";
 
 const ButtonGuide = () => {
   return (
